ENH/MAINT: implement functional.bregman.
Move the BregmanDistance functional from default_functionals.py
to functional.py, and add the function bregman to the Functional
base class.
